backend/users2/permissions.py
from rest_framework import permissions


class UserViewSetPermissions(permissions.BasePermission):
    def has_permission(self, request, view):
        if request.user.is_authenticated:
            return True
        