backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Mattéo <mattbercegeay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.0"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "42.0.5"
defusedxml = "0.7.1"
django = "4.1"
django-allauth = "0.61.1"
django-cors-headers = "4.3.1"
django-enum = "1.3.1"
django-filter = "23.5"
django-rest-auth = "0.9.5"
djangorestframework = "3.15.0"
djangorestframework-simplejwt = "5.3.1"
drf-standardized-errors = "0.13.0"
drf-yasg = "1.21.7"
gunicorn = "21.2.0"
idna = "3.6"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.3"
markdown = "3.6"
markupsafe = "2.1.5"
oauthlib = "3.2.2"
packaging = "24.0"
pillow = "10.2.0"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dotenv = "1.0.1"
python3-openid = "3.2.0"
pytz = "2024.1"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-oauthlib = "1.4.0"
six = "1.16.0"
sqlparse = "0.4.4"
typing-extensions = "4.10.0"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
whitenoise = "6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
